Unify hpl and intel-hpl via base_application
Tested with the following (checked both hpl and intel-hpl even though
only hpl is shown.) The only real difference is the addition of
`workload_group`.

```
$ ramble info -vvv hpl > new_dump.txt
$ git stash
$ ramble info -vvv hpl > old_dump.txt
$ diff old_dump.txt new_dump.txt
249a250,254
> ###################
> # workload_groups #
> ###################
> {'standard'}
> {'calculator'}
```
